test/mochitest/test_devices.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container"></div>
<pre id="test">
  <script type="application/javascript" id="fake-adb"></script>
  <script type="application/javascript">

  "use strict";

  var fakeAdbDevice, fakeFastbootDevice;

  function populateFakeDevices() {
    return loadFakeDevice().then(() => {
      fakeAdbDevice = new FakeAdbDevice();
      fakeFastbootDevice = new FakeFastbootDevice();
    });
  }

  function test_plugAdbDevice() {
    return new Promise((resolve, reject) => {
      w.Devices.register(fakeAdbDevice._serial, fakeAdbDevice);
      is(Object.keys(w.Devices._devices).length, 1, "Devices._devices has one element");

      let dev = w.Devices._devices[fakeAdbDevice._serial];
      ok(dev, "FakeAdbDevice", fakeAdbDevice._serial, "exists");
      is(dev.type, "adb", "FakeAdbDevice is ADB");
      ok(('pull' in dev), "FakeAdbDevice has adb pull");
      ok(('shell' in dev), "FakeAdbDevice has adb shell");
      ok(('rebootBootloader' in dev), "FakeAdbDevice has adb reboot bootloader");
      ok(('summonRoot' in dev), "FakeAdbDevice has adb root");

      setTimeout(() => resolve());
    });
  }

  function test_assert_device() {
    return new Promise((resolve, reject) => {
      let deviceId = c.getElementById("deviceId");
      is(deviceId.textContent, "Mochitest Fake Device", "Device detected in ADB mode");

      let unsupportedDeviceName = c.querySelector("#noDevice h4");
      is(unsupportedDeviceName, null, "No unsupported device message");
      resolve();
    });
  }

  function test_assert_builds() {
    return new Promise((resolve, reject) => {
      let devices = c.getElementById("devices");
      ok(devices, "has devices node");
      is(devices.childNodes.length, 2, "two builds available");

      let url, name, desc;
      url  = devices.getElementsByTagName("input")[0].value;
      name = devices.getElementsByTagName("h4")[0].textContent;
      desc = devices.getElementsByTagName("span")[0].textContent;
      is(url,  "http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/fake.nightly-userdebug.zip", "has userdebug url");
      is(name, "nightly userdebug", "has userdebug name");
      is(desc, "Nightly userdebug build freshly baked from TaskCluster", "has userdebug description");

      url  = devices.getElementsByTagName("input")[1].value;
      name = devices.getElementsByTagName("h4")[1].textContent;
      desc = devices.getElementsByTagName("span")[1].textContent;
      is(url,  "http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/fake.nightly-eng.zip", "has engineering url");
      is(name, "nightly eng", "has engineering name");
      is(desc, "Nightly eng build freshly baked from TaskCluster", "has engineering description");

      resolve();
    });
  }

  function test_unplugAdbDevice() {
    return new Promise((resolve, reject) => {
      w.Devices.unregister(fakeAdbDevice._serial);
      is(Object.keys(w.Devices._devices).length, 0, "Devices._devices has no element");

      resolve();
    });
  }

  function test_plugFastbootDevice() {
    return new Promise((resolve, reject) => {
      w.Devices.register(fakeFastbootDevice._serial, fakeFastbootDevice);
      is(Object.keys(w.Devices._devices).length, 1, "Devices._devices has one element");

      let dev = w.Devices._devices[fakeFastbootDevice._serial];
      ok(dev, "FakeFastbootDevice", fakeFastbootDevice._serial, "exists");
      is(dev.type, "fastboot", "FakeFastbootDevice is Fastboot");
      ok(('getvar' in dev), "FakeFastbootDevice has fastboot getvar");
      ok(('reboot' in dev), "FakeFastbootDevice has fastboot reboot");
      ok(('flash' in dev), "FakeFastbootDevice has fastboot flash");

      setTimeout(() => resolve());
    });
  }

  function test_unplugFastbootDevice() {
    return new Promise((resolve, reject) => {
      w.Devices.unregister(fakeFastbootDevice._serial);
      is(Object.keys(w.Devices._devices).length, 0, "Devices._devices has no element");

      resolve();
    });
  }

  function runTest() {
    populateAboutB2GInstaller()
      .then(() => populateFakeDevices())
      .then(() => ok(true, "testing with ADB"))
      .then(() => test_plugAdbDevice())
      .then(() => test_assert_device())
      .then(() => test_assert_builds())
      .then(() => test_unplugAdbDevice())
      .then(() => ok(true, "testing with Fastboot"))
      .then(() => test_plugFastbootDevice())
      .then(() => test_unplugFastbootDevice())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
